+++ cocos/renderer/ccShader_CameraClear.frag
@@ -2,8 +2,8 @@
 const char* ccCameraClearFrag = R"(
 
 #ifdef GL_ES
-varying mediump vec2 v_texCoord;
-varying mediump vec3 v_color;
+varying highp vec2 v_texCoord;
+varying lowp vec3 v_color;
 #else
 varying vec2 v_texCoord;
 varying vec3 v_color;
