+++ cocos/renderer/ccShader_CameraClear.vert
@@ -7,8 +7,8 @@ attribute vec4 a_position;
 attribute vec2 a_texCoord;
 attribute vec3 a_color;
 #ifdef GL_ES
-varying mediump vec2 v_texCoord;
-varying mediump vec3 v_color;
+varying highp vec2 v_texCoord;
+varying lowp vec3 v_color;
 #else
 varying vec2 v_texCoord;
 varying vec3 v_color;
