+++ cocos/renderer/ccShader_Label.vert
@@ -31,7 +31,7 @@ attribute vec4 a_color;
 
 #ifdef GL_ES
 varying lowp vec4 v_fragmentColor;
-varying mediump vec2 v_texCoord;
+varying highp vec2 v_texCoord;
 #else
 varying vec4 v_fragmentColor;
 varying vec2 v_texCoord;
