+++ cocos/renderer/ccShader_PositionTexture.vert
@@ -29,7 +29,7 @@ attribute vec4 a_position;
 attribute vec2 a_texCoord;
 
 #ifdef GL_ES
-varying mediump vec2 v_texCoord;
+varying highp vec2 v_texCoord;
 #else
 varying vec2 v_texCoord;
 #endif
